requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.0

# Groq AI
groq==0.4.0

# Video/Social media scraping
yt-dlp==2023.11.16
instaloader==4.10.3

# OCR (optional)
Pillow==10.1.0
pytesseract==0.3.10

# Utilities
pydantic==2.5.0